requirements.txt
discord.py>=2.0
google-auth>=2.0
google-api-python-client>=2.0
pytest>=7.0
pytest-asyncio>=0.21
vcrpy>=8.0
